pyproject.toml
[project]
name = "pyqres"
version = "0.1.0"
description = "Quantum Resource Estimation DSL tools"
authors = [
    { name="Agony", email="chenzhaoyun@iai.ustc.edu.cn" }
]

dependencies = [
    "numpy>=1.20.0",
    "lark",
    "sympy",
    "pyyaml",
    "pysparq",
]

[project.optional-dependencies]
test = ["pytest"]
dev = ["jupyter"]

[project.scripts]
pyqres = "pyqres.cli:main"

[tool.setuptools.packages.find]
include = ["pyqres", "pyqres.*"]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.coverage.run]
source = ["pyqres"]
